Fix TypeError by removing redundant tf.int32 cast before jax.device_put

+++ VishwamAI/scripts/train_vishwamai_model.py
@@ -74,9 +74,9 @@ def train_step(params, transformed_forward, optimizer, batch, labels, rng):
     batch = tf.cast(batch, tf.int32)
     labels = tf.cast(labels, tf.int32)
 
-    # Convert TensorFlow tensors to numpy arrays outside of the JAX-traced loss_fn
-    batch_jax = jax.device_put(tf.cast(batch, tf.int32))
-    labels_jax = jax.device_put(tf.cast(labels, tf.int32))
+    # Convert TensorFlow tensors to JAX arrays using JAX's data type specification
+    batch_jax = jax.device_put(batch)
+    labels_jax = jax.device_put(labels)
 
     # Use gradient checkpointing to save memory during the backward pass
     loss, grads = jax.value_and_grad(jax.checkpoint(loss_fn))(params)
